refactor(status): integrate custom error handling in `show_status` handler

+++ src/models/status.rs
@@ -7,8 +7,8 @@ type Version = String;
 #[derive(Deserialize, Serialize, ToSchema)]
 pub struct Database {
     pub version: Version,
-    pub max_connections: u16,
-    pub opened_connections: u16,
+    pub max_connections: i64,
+    pub opened_connections: i64,
 }
 
 #[derive(Deserialize, Serialize, ToSchema)]
